move updating to domain service, add privilege checks to add/update/delete domain service methods

+++ src/Domain.LinnApps/StockLocators/StockLocatorService.cs
@@ -3,6 +3,7 @@
     using System.Collections.Generic;
     using System.Linq;
 
+    using Linn.Common.Authorisation;
     using Linn.Common.Persistence;
     using Linn.Stores.Domain.LinnApps.Exceptions;
 
@@ -14,26 +15,48 @@
 
         private readonly IQueryRepository<StoragePlace> storagePlaceRepository;
 
+        private readonly IAuthorisationService authService;
+
         public StockLocatorService(
             IRepository<StockLocator, int> repository,
             IStoresPalletRepository palletRepository,
-            IQueryRepository<StoragePlace> storagePlaceRepository)
+            IQueryRepository<StoragePlace> storagePlaceRepository,
+            IAuthorisationService authService)
         {
             this.repository = repository;
             this.palletRepository = palletRepository;
             this.storagePlaceRepository = storagePlaceRepository;
+            this.authService = authService;
         }
 
-        public void UpdateStockLocator(StockLocator @from, StockLocator to)
+        public void UpdateStockLocator(StockLocator @from, StockLocator to, IEnumerable<string> privileges)
         {
-            throw new System.NotImplementedException();
+            if (!this.authService.HasPermissionFor(AuthorisedAction.UpdateStockLocator, privileges))
+            {
+                throw new StockLocatorException("You are not authorised to update.");
+            }
+
+            from.BatchRef = to.BatchRef;
+            from.StockRotationDate = to.StockRotationDate;
+            from.Quantity = to.Quantity;
+            from.Remarks = to.Remarks;
+            from.PalletNumber = to.PalletNumber;
+            from.LocationId = to.LocationId;
         }
 
-        public StockLocator CreateStockLocator(StockLocator toCreate, string auditDepartmentCode)
+        public StockLocator CreateStockLocator(
+            StockLocator toCreate, 
+            string auditDepartmentCode, 
+            IEnumerable<string> privileges)
         {
+            if (!this.authService.HasPermissionFor(AuthorisedAction.CreateStockLocator, privileges))
+            {
+                throw new StockLocatorException("You are not authorised to create.");
+            }
+
             if (toCreate.LocationId.HasValue == toCreate.PalletNumber.HasValue)
             {
-                throw new CreateStockLocatorException("Must Supply EITHER Location Id OR Pallet Number");
+                throw new StockLocatorException("Must Supply EITHER Location Id OR Pallet Number");
             }
 
             if (toCreate.PalletNumber != null)
@@ -43,7 +66,7 @@
                 {
                     if (storesPallet.AuditedByDepartmentCode == null && auditDepartmentCode == null)
                     {
-                        throw new CreateStockLocatorException("Audit department must be entered");
+                        throw new StockLocatorException("Audit department must be entered");
                     }
 
                     if (auditDepartmentCode != null)
@@ -66,8 +89,13 @@
             return toCreate;
         }
 
-        public void DeleteStockLocator(StockLocator toDelete)
+        public void DeleteStockLocator(StockLocator toDelete, IEnumerable<string> privileges)
         {
+            if (!this.authService.HasPermissionFor(AuthorisedAction.CreateStockLocator, privileges))
+            {
+                throw new StockLocatorException("You are not authorised to delete.");
+            }
+
             this.repository.Remove(toDelete);
             if (!this.repository
                     .FilterBy(l => l.PalletNumber == toDelete.PalletNumber && l.Quantity > 0).Any())
